refactor(core): split governance_handlers — extract protocol module
Extract protocol change handlers (handle_protocol_upgrade,
handle_protocol_change, handle_delayed_protocol_change,
emit_parameter_changed) into governance_handlers/protocol.rs.

Reduces mod.rs from ~1500 to ~1230 lines. No behavioral changes.

Ratchet: 83 → 85 (+2 from submodule use statements across federation,
treasury, and protocol splits).

Part of #859
